+++ PureEdgeSim/com/mechalikh/pureedgesim/LocationManager/Mobility.java
@@ -3,7 +3,8 @@ package com.mechalikh.pureedgesim.LocationManager;
 public abstract class Mobility {
 
 	protected Location currentLocation;
-
+	protected boolean isMobile = false;
+	
 	public Mobility(Location location) {
 		this.currentLocation = location;
 	}
@@ -15,5 +16,13 @@ public abstract class Mobility {
 
 	public Location getCurrentLocation() {
 		return currentLocation;
+	}
+	
+	public boolean isMobile() {
+		return isMobile;
+	}
+
+	public void setMobile(boolean mobile) {
+		isMobile = mobile;
 	}
 }