+++ PureEdgeSim/com/mechalikh/pureedgesim/TasksOrchestration/Orchestrator.java
@@ -117,8 +117,8 @@ public abstract class Orchestrator {
 		if (device2.getType() == TYPES.CLOUD)
 			return true;
 		double distance = Math
-				.abs(Math.sqrt(Math.pow((device1.getLocation().getXPos() - device2.getLocation().getXPos()), 2)
-						+ Math.pow((device1.getLocation().getYPos() - device2.getLocation().getYPos()), 2)));
+				.abs(Math.sqrt(Math.pow((device1.getMobilityManager().getCurrentLocation().getXPos() - device2.getMobilityManager().getCurrentLocation().getXPos()), 2)
+						+ Math.pow((device1.getMobilityManager().getCurrentLocation().getYPos() - device2.getMobilityManager().getCurrentLocation().getYPos()), 2)));
 
 		return (distance < RANGE);
 	}